pyproject.toml
[project]
name = "dnd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
]
